Fix entering wrong frame block after yielding.
The fix was to leave the frame after an exception occurred.
Previously the frames were preserved in the stack even after a
prematurely exit due to an exception. So, the next time frameEnter
was called, it would use the old frames which would cause frame's code
to be inserted in random places. The randomness was due to the use of
Date.now() in Sk.yield.